telegram-translit-foreground-sender/requirements.txt
pynput
clipboard
telethon
python-dotenv